Add files via upload
A machine learning model that can predict crop yield and market demand would use historical data on weather, soil fertility, humidity, temperature, pH, and nutrient content. Data from weather stations, soil sensors, and market reports would be gathered, cleaned, preprocessed, and split into training and testing sets. Various machine learning algorithms like decision trees, random forests, or neural networks would be applied to build a predictive model that would learn the patterns and relationships between variables and make predictions on future crop yield and market demand. Paired with a user-friendly interface, this model would allow farmers and agricultural businesses to make informed decisions about crop selection, planting strategies, and market forecasting, leading to increased crop yields and profits. 

As of now, we are working on developing the front end, but the machine learning model has been trained and recommends the crops to be grown according to the input values.
We have put in standard input values,they can be changed to see variation in the o/p.
The o/p displays a list of crops from best to worst to be grown according to the i/p.

The trained model has been loaded already, so to run this project :
Run the main.py file, the o/p will be visible in the terminal.